[#20] Refactor: helpers 폴더 하위에 chart 폴더 생성 및 차트 데이터 분리
