+++ src/pages/Main.tsx
@@ -1,7 +1,7 @@
 import { Container } from '@mui/system';
 import { useLoaderData, useLocation } from 'react-router-dom';
 
-import type { MockData } from '@/types/mockData';
+import type { MockDataType } from '@/types/mockData';
 import type { QueryDataType } from '@/types/queryData';
 
 import FilterableChart from '@/components/FilterableChart';
@@ -10,7 +10,7 @@ import getQueryData from '@/helpers/getQueryData';
 
 const Main = () => {
   const location = useLocation();
-  const loaderData = useLoaderData() as MockData;
+  const loaderData = useLoaderData() as MockDataType;
   const queryData = getQueryData(location.search);
 
   return (
